When formatting debug messages with int64_t and uint64_t values, we have to take cautions to not use explicitly the %li and %lu format specifiers because theint64_t and uint64_t types are alias to the more native types encoded in 64 bits and those native types depend on the underlying physical processor architeture. For example, uint64_t is an unsigned long on Intel/AMD architectures but it's an unsigned long long on Apple ARM64 architectures. So, there is two ways to circumvent this problem: - either cast the value to the type expected by the format specifier (for example, for %lu, cast the value as an unsigned long) - or use the macro defined in inittypes.h to specify the correct format specifier for int64_t and uint64_t values for the underlying processor architecture.
The later is used in this fix. For example, replace the statement

`dt_print(DT_DEBUG_PIPE, "couldn't allocate new cache entry %lu\n", hash);`

by the following one

`dt_print(DT_DEBUG_PIPE, "couldn't allocate new cache entry %" PRIu64 "\n", hash);`

In other terms, please use PRIu64 for uint64_t and PRId64 for int64_t.

+++ src/develop/imageop.c
@@ -1774,7 +1774,7 @@ void dt_iop_commit_params(dt_iop_module_t *module, dt_iop_params_t *params,
   piece->global_hash = piece->hash = hash;
   piece->global_mask_hash = piece->blendop_hash = module->blendop_hash;
 
-  dt_print(DT_DEBUG_PIPE, "[pixelpipe] params commit for %s (%s) in pipe %i with hash %lu\n", module->op, module->multi_name, pipe->type, (long unsigned int)piece->hash);
+  dt_print(DT_DEBUG_PIPE, "[pixelpipe] params commit for %s (%s) in pipe %i with hash %" PRIu64 "\n", module->op, module->multi_name, pipe->type, piece->hash);
 }
 
 void dt_iop_gui_cleanup_module(dt_iop_module_t *module)
